refactor: extract split-layout styles to reusable css literal
